mv Accesskey into auth_ext.go
