--- filter/accesskey.go
@@ -1,10 +1,0 @@
-package filter
-
-type AccessKeyPair struct {
-	AccessKey    string
-	SecretKey    string
-	ConsumerSide string
-	ProviderSide string
-	Creator      string
-	Options      string
-}
